Moved updated find_att function to utils/common.py
