+++ medallion/utils/common.py
@@ -148,3 +148,24 @@ def generate_status_details(id, version, message=None):
         status_details["message"] = message
 
     return status_details
+
+
+def find_att(obj):
+    """
+    Used for finding the version attribute of an ambiguous object. Manifests
+    use the "version" field, but objects will use "modified", or if that's not
+    available, the "created" field.
+    Args:
+        obj (dict): manifest or stix object
+    Returns:
+        string value of the field from the object to use for versioning
+    """
+    if "version" in obj:
+        return convert_to_stix_datetime(obj["version"])
+    elif "modified" in obj:
+        return convert_to_stix_datetime(obj["modified"])
+    elif "created" in obj:
+        return convert_to_stix_datetime(obj["created"])
+    else:
+        # TO DO: PUT DEFAULT VALUE HERE
+        pass
